pyproject.toml
[tool.poetry]
name = "iprovide"
version = "0.1.0"
description = "Provides a simple but effective  interface definition library."
authors = ["David James McCorrie <djmccorrie@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyhamcrest = "^2.0.2"
pylint = "^2.5.3"
mypy = "^0.782"
pdbpp = "^0.10.2"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
